Fix consutructor issue and add file path check

+++ tests/test_jsondbupload.py
@@ -2,7 +2,8 @@ import unittest, os, sys
 sys.path.insert(0, '../jsondbupload/')
 from jsondbupload import JsonDBUpload
 
-from zohavi_base_model.models import TableDefault
+from zohavi.zdb.models import TBL_Default
+
 from flask import Flask
 from flask_sqlalchemy import SQLAlchemy, Model
 
@@ -11,12 +12,12 @@ from mclogger import MCLogger
 #################################################################################
 class Models_Book: 
 	def __init__(self, db):
-		class Author(db.Model, TableDefault):
+		class Author(db.Model, TBL_Default):
 			__tablename__ = 'author' 
 			id = db.Column(db.Integer(), primary_key=True)
 			name = db.Column(db.Integer() )
 
-		class Book(db.Model, TableDefault):
+		class Book(db.Model, TBL_Default):
 			__tablename__ = 'book' 
 			id = db.Column(db.Integer(), primary_key=True)
 			name = db.Column(db.Integer() )
@@ -24,12 +25,12 @@ class Models_Book:
 			_author = db.relationship("Author", backref=db.backref("author" ), lazy='joined')
 
 
-		class Bookset(db.Model, TableDefault):
+		class Bookset(db.Model, TBL_Default):
 			__tablename__ = 'bookset' 
 			id = db.Column(db.Integer(), primary_key=True)
 			name = db.Column(db.Integer() )
 			
-		class BooksetItems(db.Model, TableDefault):
+		class BooksetItems(db.Model, TBL_Default):
 			__tablename__ = 'bookset_items' 
 			id = db.Column(db.Integer(), primary_key=True)
 			bookset_id = db.Column( db.Integer()  , db.ForeignKey( 'bookset.id'  ) )
@@ -44,6 +45,29 @@ class Models_Book:
 
 		self.tables_dict = {table.__tablename__: table for table in db.Model.__subclasses__()}
 
+class Models_Config: 
+	def __init__(self, db):
+		class TBL_Config(db.Model, TBL_Default):
+			__tablename__ = 'config' 
+			id = db.Column(db.Integer(), primary_key=True)
+			area 		= db.Column(db.String(25) ) 
+			category 	= db.Column(db.String(25) ) 
+			sub_cat		= db.Column(db.String(25) ) 
+
+		class TBL_ConfigItem(db.Model, TBL_Default):
+			__tablename__ = 'config_item' 
+			id = db.Column(db.Integer(), primary_key=True)
+			config_id = db.Column(db.Integer(), db.ForeignKey( 'config.id'  ) )
+			_config = db.relationship("TBL_Config", backref=db.backref("config" ), lazy='joined')
+			env =  db.Column(db.String(10) ) 
+			name = db.Column(db.String(100) ) 
+			value = db.Column(db.String(255) ) 		
+
+		self.TBL_Config = TBL_Config
+		self.TBL_ConfigItem = TBL_ConfigItem
+
+		self.tables_dict = {table.__tablename__: table for table in db.Model.__subclasses__()}
+
 	#################################################################################
 	def table_object(table_name):
 		return self.tables_dict.get(table_name)
@@ -51,7 +75,7 @@ class Models_Book:
 #################################################################################
 class TestJsonDBUpload(unittest.TestCase):
 	def _db_session(self):
-		os.remove('test.db')
+		if os.path.exists('test.db'): os.remove('test.db')
 		app = Flask(__name__)
 		app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///test.db'
 		app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
@@ -62,7 +86,8 @@ class TestJsonDBUpload(unittest.TestCase):
 	#################################################################################
 	def _db_create_tables(self, db):
 		# db._models = {}
-		db._models = Models_Book(db)
+		db._models1 = Models_Book(db)
+		db._models2 = Models_Config(db)
 		db.create_all()
 
 	#################################################################################
@@ -73,16 +98,22 @@ class TestJsonDBUpload(unittest.TestCase):
 	#################################################################################
 	def test_insert_recort(self):
 		db = self._db_session()
-		auth = db._models.Author(name='John Smith')
+		auth = db._models1.Author(name='John Smith')
 		db.session.add(auth)
 		db.session.commit()
 		self.assertEqual( True, isinstance(auth.id, int))
 	
 	#################################################################################
-	def _create_json_to_db_instance(self, db, db_upload_dict):
+	def _create_json_to_db_instance_from_dict(self, db, db_upload_dict):
 		logger = MCLogger( 'test_log.text')
 		j2db = JsonDBUpload( db, logger )
 		j2db.update_tables_from_dict( db_upload_dict) 
+
+	#################################################################################
+	def _create_json_to_db_instance_from_file(self, db, db_file_name):
+		logger = MCLogger( 'test_log.text')
+		j2db = JsonDBUpload( db, logger )
+		j2db.update_tables_from_file( db_file_name) 
 
 	#################################################################################
 	def test_json_insert_single(self):
@@ -97,9 +128,9 @@ class TestJsonDBUpload(unittest.TestCase):
 					}
 		]
 		db = self._db_session()
-		j2db = self._create_json_to_db_instance( db, db_upload_dict )
+		j2db = self._create_json_to_db_instance_from_dict( db, db_upload_dict )
 
-		auth_list = db.session.query( db._models.Author ).all()
+		auth_list = db.session.query( db._models1.Author ).all()
 
 		for index, item in enumerate(auth_list):
 			self.assertEqual( auth_list[ index ].name, db_upload_dict[0]['data'][index ]['name'])
@@ -127,10 +158,10 @@ class TestJsonDBUpload(unittest.TestCase):
 					
 		]
 		db = self._db_session()
-		j2db = self._create_json_to_db_instance( db, db_upload_dict )
+		j2db = self._create_json_to_db_instance_from_dict( db, db_upload_dict )
 
 		# auth_list = db.session.query( db._models.Author ).all()
-		book_list = db.session.query( db._models.Book ).all()
+		book_list = db.session.query( db._models1.Book ).all()
 
 		for index, item in enumerate(book_list):
 			# breakpoint()
@@ -176,10 +207,10 @@ class TestJsonDBUpload(unittest.TestCase):
 					
 		]
 		db = self._db_session()
-		j2db = self._create_json_to_db_instance( db, db_upload_dict )
+		j2db = self._create_json_to_db_instance_from_dict( db, db_upload_dict )
 
 		# auth_list = db.session.query( db._models.Author ).all()
-		booksets = db.session.query( db._models.BooksetItems ).all()
+		booksets = db.session.query( db._models1.BooksetItems ).all()
 
 		for index, item in enumerate(booksets):
 			# breakpoint()
@@ -187,7 +218,19 @@ class TestJsonDBUpload(unittest.TestCase):
 			self.assertEqual( booksets[ index ]._book.name    , db_upload_dict[1]['data'][index ]['name'])
 	
 	
+	#################################################################################
+	def test_json_insert_from_file(self):
+		db = self._db_session()
+		j2db = self._create_json_to_db_instance_from_file( db, 'tests/db_config_defaults.json' )
 
+		# auth_list = db.session.query( db._models.Author ).all()
+		# booksets = db.session.query( db._models.BooksetItems ).all()
+
+		# for index, item in enumerate(booksets):
+		# 	# breakpoint()
+		# 	self.assertEqual( booksets[ index ]._bookset.name , db_upload_dict[2]['data'][0 ]['name'])
+		# 	self.assertEqual( booksets[ index ]._book.name    , db_upload_dict[1]['data'][index ]['name'])
+	
 
 if __name__ == '__main__':
     unittest.main()
